move slash handler logic back into slash commands
